Split up TestWithMiniCluster
