+++ src/test/java/com/paypal/nnanalytics/TestWithMiniClusterBase.java
@@ -25,7 +25,6 @@ import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.StringContains.containsString;
 
 import com.paypal.namenode.WebServerMain;
-import com.paypal.security.SecurityConfiguration;
 import java.io.File;
 import java.io.IOException;
 import java.util.Map;
@@ -33,17 +32,14 @@ import java.util.concurrent.TimeUnit;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.CommonConfigurationKeys;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.hdfs.DFSConfigKeys;
 import org.apache.hadoop.hdfs.DFSTestUtil;
 import org.apache.hadoop.hdfs.DistributedFileSystem;
 import org.apache.hadoop.hdfs.HdfsConfiguration;
 import org.apache.hadoop.hdfs.MiniDFSCluster;
 import org.apache.hadoop.hdfs.qjournal.MiniQJMHACluster;
 import org.apache.hadoop.hdfs.server.namenode.SRandom;
-import org.apache.hadoop.hdfs.server.namenode.ha.HATestUtil;
 import org.apache.http.HttpHost;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
@@ -51,76 +47,25 @@ import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.DefaultHttpClient;
 import org.junit.AfterClass;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Ignore;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
 
-@RunWith(JUnit4.class)
-public class TestWithMiniCluster {
+public abstract class TestWithMiniClusterBase {
 
-  private static final SRandom RANDOM = new SRandom();
-  private static final int NUMDATANODES = 1;
-  private static final Configuration CONF = new HdfsConfiguration();
-  private static final String NAMESERVICE = MiniQJMHACluster.NAMESERVICE;
+  protected static final SRandom RANDOM = new SRandom();
+  protected static final int NUMDATANODES = 1;
+  protected static final Configuration CONF = new HdfsConfiguration();
+  protected static final String NAMESERVICE = MiniQJMHACluster.NAMESERVICE;
 
-  private static final byte[] TINY_FILE_BYTES = new byte[512];
-  private static final byte[] SMALL_FILE_BYTES = new byte[1024];
-  private static final byte[] MEDIUM_FILE_BYTES = new byte[1024 * 1024];
-  private static final String[] USERS = new String[] {"hdfs", "test_user"};
+  protected static final byte[] TINY_FILE_BYTES = new byte[512];
+  protected static final byte[] SMALL_FILE_BYTES = new byte[1024];
+  protected static final byte[] MEDIUM_FILE_BYTES = new byte[1024 * 1024];
+  protected static final String[] USERS = new String[] {"hdfs", "test_user"};
 
-  private static MiniQJMHACluster cluster;
-  private static HttpHost hostPort;
-  private static HttpClient client;
-  private static WebServerMain nna;
-
-  @BeforeClass
-  public static void beforeClass() throws Exception {
-    RANDOM.nextBytes(TINY_FILE_BYTES);
-    RANDOM.nextBytes(SMALL_FILE_BYTES);
-    RANDOM.nextBytes(MEDIUM_FILE_BYTES);
-
-    // Speed up editlog tailing.
-    CONF.setInt(DFSConfigKeys.DFS_HA_TAILEDITS_PERIOD_KEY, 1);
-    CONF.setInt(DFSConfigKeys.DFS_HA_LOGROLL_PERIOD_KEY, 1);
-    CONF.setInt(DFSConfigKeys.DFS_DATANODE_SCAN_PERIOD_HOURS_KEY, -1);
-    CONF.setInt(DFSConfigKeys.DFS_CLIENT_RETRY_WINDOW_BASE, 10);
-    CONF.setBoolean("fs.hdfs.impl.disable.cache", true);
-    CONF.set(CommonConfigurationKeys.FS_DEFAULT_NAME_KEY, "hdfs://" + NAMESERVICE);
-
-    MiniQJMHACluster.Builder qjmBuilder = new MiniQJMHACluster.Builder(CONF);
-    qjmBuilder.getDfsBuilder().numDataNodes(NUMDATANODES);
-    cluster = qjmBuilder.build();
-    cluster.getDfsCluster().waitActive();
-    cluster.getDfsCluster().transitionToActive(0);
-
-    HATestUtil.setFailoverConfigurations(cluster.getDfsCluster(), CONF, NAMESERVICE, 0);
-    CONF.set("dfs.nameservice.id", NAMESERVICE);
-
-    nna = new WebServerMain();
-    SecurityConfiguration nnaConf = new SecurityConfiguration();
-    nnaConf.set("nna.support.bootstrap.overrides", "true");
-    nnaConf.set("ldap.enable", "false");
-    nnaConf.set("authorization.enable", "false");
-    nnaConf.set("nna.historical", "true");
-    nnaConf.set("nna.base.dir", MiniDFSCluster.getBaseDirectory());
-    nna.init(nnaConf, null, CONF);
-    hostPort = new HttpHost("localhost", 4567);
-    client = new DefaultHttpClient();
-
-    // Fetch NNA Namespace.
-    HttpGet fetch = new HttpGet("http://localhost:4567/fetchNamespace");
-    HttpResponse fetchRes = client.execute(hostPort, fetch);
-    assertThat(fetchRes.getStatusLine().getStatusCode(), is(200));
-    IOUtils.readLines(fetchRes.getEntity().getContent());
-
-    // Reload NNA Namespace.
-    HttpGet reload = new HttpGet("http://localhost:4567/reloadNamespace");
-    HttpResponse reloadRes = client.execute(hostPort, reload);
-    assertThat(reloadRes.getStatusLine().getStatusCode(), is(200));
-    IOUtils.readLines(reloadRes.getEntity().getContent());
-  }
+  protected static MiniQJMHACluster cluster;
+  protected static HttpHost hostPort;
+  protected static HttpClient client;
+  protected static WebServerMain nna;
 
   @AfterClass
   public static void tearDown() throws IOException {
@@ -135,16 +80,6 @@ public class TestWithMiniCluster {
   @Before
   public void before() throws IOException {
     client = new DefaultHttpClient();
-  }
-
-  /**
-   * Long running execution that will launch an NNA instance backed by a live updating HA-enabled
-   * HDFS instance. NNA will update every minute or so with new analysis.
-   */
-  public static void main(String[] args) throws Exception {
-    beforeClass();
-    TestWithMiniCluster test = new TestWithMiniCluster();
-    test.addFiles(10000, 300L);
   }
 
   @Test
@@ -275,7 +210,7 @@ public class TestWithMiniCluster {
     assertThat(checkCount, is(greaterThan(startingCount)));
   }
 
-  private void addFiles(int numOfFiles, long sleepBetweenMs) throws Exception {
+  protected void addFiles(int numOfFiles, long sleepBetweenMs) throws Exception {
     DistributedFileSystem fileSystem = (DistributedFileSystem) FileSystem.get(CONF);
     for (int i = 0; i < numOfFiles; i++) {
       int dirNumber1 = RANDOM.nextInt(10);
